requirements.txt
Flask
pypdf
openai
PyYAML
google-generativeai
python-dotenv